+++ src/sources/native/buffer.ts
@@ -65,6 +65,7 @@ export default class Buffer extends Source {
   public async doComplete(opt: CompleteOption, token: CancellationToken): Promise<CompleteResult> {
     let { bufnr, input } = opt
     if (input.length == 0) return null
+    await waitImmediate()
     let words: Set<string> = new Set()
     let isIncomplete = await this.getWords(bufnr, opt, token, words)
     return {